Refactored safety page to add routing, preparation for Lost and Found pages
